move handle accept to actions
