+++ apps/share-easyi-client/src/components/call-dialog.tsx
@@ -3,7 +3,6 @@ import { callStatus, type CallStatus } from '@/call-service/store'
 import { CallApi } from '@/call-service'
 import { Button } from '@/components/ui/button'
 import { Stack } from '@/components/ui/stack'
-import { EEmit } from '@/event-bus/emitter'
 import { cn } from '@/utils'
 import { createEffect, For, Show } from 'solid-js'
 import { CallRipple } from './call-ripples'
